LECON-54 final frontend design. working on buttons
